monitoring/requirements.txt
paho-mqtt==2.1.0
requests==2.32.3
python-dotenv==1.0.1
influxdb-client==1.45.0
pydantic==2.9.2
